templates html files added

+++ src/main/java/com/birkil/hospitalmonitoring/service/HastaBilgileriService.java
@@ -2,6 +2,7 @@ package com.birkil.hospitalmonitoring.service;
 
 import com.birkil.hospitalmonitoring.entity.HastaBilgileri;
 import com.birkil.hospitalmonitoring.repository.HastaBilgileriRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,6 +13,7 @@ public class HastaBilgileriService {
 
     private final HastaBilgileriRepository hastaBilgileriRepository;
 
+    @Autowired
     public HastaBilgileriService(HastaBilgileriRepository hastaBilgileriRepository) {
         this.hastaBilgileriRepository = hastaBilgileriRepository;
     }
@@ -29,11 +31,36 @@ public class HastaBilgileriService {
         return hastaBilgileriRepository.save(hastaBilgileri);
     }
 
-    public HastaBilgileri updateHasta(HastaBilgileri hastaBilgileri) {
-        return hastaBilgileriRepository.save(hastaBilgileri);
+    public HastaBilgileri updateHasta(Long id, HastaBilgileri hastaBilgileri) {
+
+        HastaBilgileri existingHasta = getHastaById(id);
+        if(existingHasta != null){
+            existingHasta.setAd(hastaBilgileri.getAd());
+            existingHasta.setSoyad(hastaBilgileri.getSoyad());
+            existingHasta.setTcKimNo(hastaBilgileri.getTcKimNo());
+            existingHasta.setCinsiyet(hastaBilgileri.getCinsiyet());
+            existingHasta.setDogumTarihi(hastaBilgileri.getDogumTarihi());
+            existingHasta.setSigortaliTuru(hastaBilgileri.getSigortaliTuru());
+            existingHasta.setDevredilenKurum(hastaBilgileri.getDevredilenKurum());
+            existingHasta.setKatilimPayindanMuaf(existingHasta.getKatilimPayindanMuaf());
+            existingHasta.setKapsamAdi(existingHasta.getKapsamAdi());
+            existingHasta.setIlaveUcrettenMuaf(existingHasta.getIlaveUcrettenMuaf());
+            //existingHasta.setTakip(existingHasta.getTakip());
+
+            return hastaBilgileriRepository.save(existingHasta);
+        }
+        else {
+            return null;
+        }
     }
 
-    public void deleteHasta(Long id) {
-        hastaBilgileriRepository.deleteById(id);
+    public boolean deleteHasta(Long id) {
+        if (hastaBilgileriRepository.existsById(id)) {
+            hastaBilgileriRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+
     }
 }
